Good progress. Now producing the right position error gradient using the velocity error routine. The problem was that Kane's angular velocity to Euler angle derivative routines, at least for body-fixed Euler sequences, expect angular velocity of body in parent to be expressed in *body* but I was supplying it expressed in *parent*.
There is a problem in the test case with a weld constraint: udots are non-zero when the free joint orientation is at 0 0 0. When it is away from that it works fine. Is there a redundant constraint in that configuration which is not being dealt with properly in the acceleration constraint-handling code?