Modified Body, MobilizedBody, and Visualizer APIs to return a reference index for any geometry added to those objects, and added methods for getting access to those objects later given the index. This is a first installment towards a feature requested by Tim Dorn that would allow keeping track of any geometry attached to a System so that it can be modified during simulation, for example to change the color of a leg at foot strike.
This is a potentially breaking change to the API because addDecoration()-like methods now return an int where in some cases they were previously returning a reference to the underlying object. I did not find anywhere in OpenSim that made use of those return values, but some of the Simbody examples and tests did and I fixed them in this checkin.