First pass at moving from PDF to Markdown/HTML.
Moving build-from-source documentation to Markdown, that gets picked up
by Doxygen. A user who is about to build from source can learn how to do
so either by reading the Markdown file, or reading the Doxygen output
online, or downloading a pre-built version of the library and viewing
the Doxygen documentation locally.