Create BodyFollower FrameController, rework RollingDisc.
I wanted to be able to follow the disc around with the camera.
At first I wanted to use frames that made sense given the definition
of mobilizers (e.g., pins rotate around their z axis). In this commit,
I decide to instead use frames that make the visualization easier to
think about.

I've done most of the work for a BodyFollower
FrameController that can follow around a MobilizedBody.

However, when there is no damping, energy is not conserved. It seems to
be the second (side-slip) NoSlip1D constraint. If I comment that out,
energy is conserved, but the motion of the disc is certainly
unrealistic. I've tried a few things, such as using yawMB as the Case
for this second NoSlip1D, but this caused energy to decrease with time
    (whereas, before, it was increasing with time).