Comply with C++11 standard, as enforced by Clang.
Update travis to build with C++11 support.

Still need to edit compiler flags in CMake files to enforce C++11. This
might mean some platforms (Ubuntu 12.10) may not be supported well
anymore. This might be okay because a new long term support release of
Ubuntu is coming out in a few months.