Fixes #130.
I decided to leave SIMBODY_SONAME_VERSION as a CACHE variable, but to
use the FORCE setting for it. The reason is that I want to be able to
see it in the CMake GUI, but I don't actually want it to be cached. This
is what we are already doing with SIMBODY_VERSION.