Fixes #139.
If using gcc or Clang and using C++11, then we fill in the OVERRIDE_11
and FINAL_11 macros with `override`, and `final`.

Note that the way I check if we're using C++11 is not completely
legitimate. The internet suggests doing feature-by-feature checks, since
most compiliers who say they are C++11 compliant are not necessarily
fully C++11 compliant. For the current C++11 features we use, there is
no issue.