Improved Simbody's version-checking mechanism.
This requires CMake 2.8.6 (only 2.8 was required previously). We use a script
to automatically generate a SimbodyConfigVersion.cmake file. Previously, I
wrote one myself, but it was likely to be buggy.

This new script enforces the behavior that changes to the major version number
are backwards-incompatible, but changes to the minor or patch versions are
backwards-compatible.

I tested this on Linux, with OpenSim as a dependent project.