SimbodyConfig.cmake allows relocation of installation.
Previously, SimbodyConfig.cmake contained hardcoded paths to the installation
directory. This is fine on Linux, but may be an issue on Windows where users
choose the installation directory long after the build process.

To do this, I used CMake module that comes with CMake 2.8.8. We currently only
support up through CMake 2.8.6. Since Ubuntu 12.04 uses 2.8.6 and it is likely
that our users will be running Ubuntu 12.04, we cannot use CMake 2.8.8 yet.
Instead, I copied the necessary CMake module from 2.8.8 and committed it to the
repo.

I tested this on Linux by compiling a Simbody-using client project, and by
changing the name of the install directory after the installation was complete.
In both cases, my client project was able to correctly find Simbody. I do not
think a Windows test is necessary, as I did not change any platform specific
behavior.

[cmake] Include local copy of CMakePackageConfigHelpers.

[travis] Use cmake 2.8.8.

Update .travis.yml

[travis] Installing cmake 2.8.8 in /usr.

Upgrading CMake to 2.8.8.

[travis] cmake 2.8.8 shell installer.

[travis] Fix typo for cmake install.

[travis] Try not installing cmake to /usr.

[travis] Back to the bash script.

[travis] bash script.

[travis] Debugging cmake install.

[travis] Install cmake locally.

[travis] mkdir cmake-install.

[travis] mkdir cmake-install.

[travis] cmake.

[travis] cmake.

[travis] cmake.

[travis] cmake.

[travis] cmake.

[travis] cmake.