[static] Added the possibility to compile simbody in a all static way
This commit does not change anything to the previous configuration and compilation steps for simbody

This option allows to disactivate dynamic compilation.
When needing only static version, we can avoid dynamic compilation, that require to have blas/lapack to be compiled with -fPIC for static use of these versions.
This allows to use static versions of blas/lapack coming from Linux distro, that do not use -fPIC for static libraries