Update ExponentialSpringForce.cpp
I added "override" statements for the virtual functions defined in various Simbody base classes.  The Mac build complained about this, but not the Windows build.  Not sure why this error has shown up now.  Perhaps the Mac compiler / build requirements changed?  Hmmmm.   Anyway, it is good practice to have the "override", so good to have them in there now.