Update ExponentialSpringForce.h
Ok, I think I have found the build errors that are causing the Continuous Integration (CI) tests to fail for Mac and Ubuntu.  Neither the Mac or Ubuntu platforms like Unicode symbols (e.g., μₖ) in argument lists.  In argument lists, I have reverted to using mus and muk for μₛ and μₖ.  I thought using Unicode symbols in an argument list might be kind of sketchy.  Good to try and confirm that using Unicode symbols in an argument list indeed crosses over the sketchy threshold.