Updated ExponentialSpringForce.h and ExponentialSpringForce.cpp
There is now only one constructor for class ExponentialSpringForce.  To make this change, an optional default ExponentialSpringParameters object is sent in through the argument list.

Removed the upper bound on μₛ.  Both μₛ and μₖ are still required to be ≥ 0.0, and μₖ is required to be ≤ μₛ.  I updated the Doxygen comments to be consistent with this change.

I did some limited testing to verify that these changes didn't break anything.  For example, I ran a simulation with non-default parameters and also with  μₛ = 3.5 (greater than 1.0).  Everything appears to be working properly.