Updated ExponentialSpringForce.h and ExponentialSpringForce.cpp
Class ExponentialSpringData was used as a simple way for users to gain access to detailed quantities computed for an ExponentialSpringForce instance.  The access was only through a const reference and so could not be used to alter the underlying data. However, the access was granted by making the member variables of ExponentialSpringData public.  Exposing the structure of the data cache was judged to be too brittle (too easily breakable).  Therefore, ExponentialSpringData was removed from ExponentialSpringForce.h, removing the ability of users to access the data members directly.

Instead, ExponentialSpringData is now located in ExponentialSpringForce.cpp and used only by class ExponentialSpringForceImp as an internal data structure for storing and retrieving important computed quantities in the data cache.  Access to the computed quantities for the typical user is granted via a number of conventional "get" methods, which are exported in the .h file.

A number of the more desirable "get" methods have been implemented to make sure that that the access methods function properly.  There are, however, more to come.

With this commit, access to the following three quantities, which are not associated with the data cache, is also granted via "get" methods:
1. Transform specifying the Contact Plane frame.
2. MobilizedBody for which an exponential spring was instantiated.
3. Station (point) on the body that interacts with the contact plane.