Update ExponentialSpringForce.cpp
There are no functional changes in the Exponential Spring force class with this commit, but I did do a major reorganization of the file layout.  Modeling after other .cpp files in Simbody, I pulled all of the methods in ExponentialSpringForceImpl into the class declaration.  So there are no longer any method definitions to be found elswhere.  I also re-ordered the methods so that the short accessor methods came first and the methods in which most of the meaningful calculations follow at the end.  And, I also got rid of comments that I thought were unnecessary.

I ran a few simulations to verify that I didn't break anything major.  The simulations look to be performing as expected.