Update TestExponentialSpring.cpp
The unit test for class ExponentialSpringForce now has working code that tests the details of the spring force calculations, as well as conservation of energy.  I've also reworked the structure of the tests so that many tests can be applied to different simulations.

There is still a fair amount to be done to complete what I have in mind for the unit test.  I'm committing now mainly to backup my work.