Update TestExponentialSpring.cpp
Added the ability to visualize test simulations, which can be quite useful when debugging a simulation.  Visualization should be turned off before committing code so that visualization windows do not pop up and hang regression tests.  To turn all visualization off, the static variable VizOff should be set to true.