Update ExponentialSpringForce.h
I revised the Doxygen comments to reflect the change from using the y-axis of the contact plane to using the z-axis to define the surface normal.

The normal force is along the positive z-axis of the contact plane.
The friction force lies in a plane that is parallel to the xy-plane of the contact plane.