Update ExponentialSpringsComparison.cpp
A few minor tweaks to some parameters and minor changes to the reporters.

The only significant change is the adjustment needed to make the ZAxis of the contact plane point in the direction of the normal, as opposed to the YAxis.  The adjustment required a -90.0 deg rotation about the XAxis.