Update ExponentialSpringParameters.cpp
I changed vSettle from 0.01 m/s to 0.001 m/s. With this change, the final sizes of the friction force was at least a factor of 10 smaller by the time a body came to rest.  For example, instead of a friction force of 1.6 Newtons, the final friction force was 0.12 Newtons.  The normal force in this example was ~25 Newtons.