Bug Fix related to enforcing bounds on the normal force.
One of the tests executed by TestExponentialSpring was failing when bounds were enforced on the normal force (fz). The test failed because when fz was clamped to be above 0.0 an below 100000.0, the values of fzElas and fzDamp were not made consistent with the new value of fz.

Now, the equation fz = fzElas + fzDamp should always hold true.

Indeed, a test that was failing (testTumblingWithDampWithFric) because of this bug is now passing.