Added a parameter for the settle acceleration.
The condition for a spring to transition from a sliding state (Sliding = 1.0) to a fixed state (Sliding = 0.0) is now essentially static equilibrium.  By default, in addition to the velocity having to be small to transition to a fixed state, the acceleration must now be small as well.

The default value for the settle acceleration is 0.1m/s².  The settle acceleration can be altered by calling ExponentialSpringParameters::setSettleAcceleration().