Added friction forces due to the spring model to the data cache.
After the recent revisions to the friction calculations, some calculated quantities were not saved to the data cache.  These quantities were:
fricElasSpr, fricDampSpr, and fricSpr, which are the friction forces due to a pure friction spring model.

These quantities are now saved to the data cache.  At present, there is no way to retrieve them, but accessors could be written should there be a desire to retrieve them.