Added option to switch friction models in ExponentialSpringForceImpl.
The latest friction model is the blended model, one that merges a pure damping model with a linear spring model.  There may be a few situations in which the blended model is not superior to the spring only model.  More investigation is needed.  I therefore added a flag (useBlended) in ExponentialSpringForceImpl that switches between the two calculation approaches.