Cleaned up the force calculation code.
Previously, the spring calculations were all performed in a single method (realizeSubsystemDynamicsImpl()), which got long.  In addition, I wanted to save code for alternate friction models for further investigation.  The best way to do this was to section off parts of the force calculations into different subroutines, which is now done.

I believe the code is now more readable. I tested the code to make sure that it didn't run any differently.