Removed make_unique<> from the code.
It looks like the problem was no related to having the right header file. Maybe make_unique<> was not yet included in the C++ standard on Mac and Ubuntu.

Should I set a flag for a later version of C++?

As a potential solution, I replaced instances of make_unique<> with a longer more cumbersome call to unique_ptr<>.