Bug Fix: getSpringZeroPosition() now returns the cached update value.
Only the cached update value of the spring zero is guaranteed to be consistent with the system State, and the cached update value is only set once the System has been realized through Stage::Dynamics.

Returning the cached update value from getSpringZeroPosition() set off a chain of unit test failures.  Correction of those errors required revisiting the "invalidates" and "DependsOn" settings for the spring zero and realizing through Stage::Dynamics before accessing the spring zero.

getSpringZeroPosition() is now returning the correct value and the unit test for ExponentialSpringForce is now passing again.