Revised ExponentialSpringParameters.h based on a review by @sherm1.
Class ExponentialSpringParameters is now using the default implementations of the copy constructor and copy assignment operator.

I also made a number of changes to method names.  In particular, setElasticity() has been changed to setFrictionElasticity() and setViscosity() has been changed to setFrictionViscosity().  The same is true for the corresponding "get" methods.

These new names are now parallel to the names of the methods for the normal force (e.g., setNormalElasticity() and setNormalViscosity()) and will be much more clear to readers of calling code.

I made the needed name adjustments in the downstream calling code. All tests are passing.