Consistently use "body station" not "spring station" in the documentation.
Prior to this commit, the phrase "spring station" was often used to refer to the point on the body at which an ExponentialSpringForce instance acts. I thought this phrase was a bit unclear because of the ambiguity of the word "spring".  Does "spring" refer generally to an ExponentialSpringForce instance or specifically to the frictional spring used in Friction Model 2? Neither really.

The term "body station" is now define in the very first Doxygen paragraph in ExponentialSpringForce.h, which allows subsequent documentation to be a bit streamlined.