Now applying the Newton's 3rd Law pair to Ground.
Prior to this commit, the force generated by ExponentialSpringForce was applied only to the MobilizedBody, which is a clear violation of Newton's 3rd Law:

"If Body A applies a force to Body B, then Body B applies a force to Body A that is equal in magnitude and opposite in direction."

With this commit, the equal and opposite force is now properly also applied to Ground.  This should generate a reaction force registered by Ground, which provides the basis for another test that should be added to TestExponentialSpring.cpp.

I also got rid of an unused member variable called useBlended.