Implemented realizePosition() and realizeVelocity() methods.
I created the Position and Velocity stage methods: realizeSubsystemPositionImpl() and realizeSubsystemVelocityImpl().

Now entries in the data cache are set in the intended way. I was able to remove the calcStationKinematics() and now accessor methods getStationPosition() and getStationVelocity() don't have to do any re-calculations; they just return what's already in the data cache.

I also cleaned up the comments for the "realize" methods.

Tests are all passing.