Added maximum normal force to ExponentialSpringParameters.
Prior to this commit, a limit on the normal force was hard coded to 100,000.0 Newtons.  While this force worked well for the default parameters, it may not work well should a user want to specify a set of custom parameters that possibly use a different system of units.

So, with this commit, the maximum allowed normal force can be set programmatically. The "100,000.0 Newtons" is no longer hard coded.

ExponentialSpringParameters now has setMaxNormalForce() and getMaxNormalForce() methods.

TestExponentialSpring.cpp has been updated to test the new maxFz parameter.

All tests are passing.