Added a test to help ensure friction is being tested properly.
At one point, I inadvertently set the static and kinetic coefficients of friction to zero, even when these constants should have been non-zero. The problematic thing was that all tests still passed.

With this commit, I added a one line test to checkSpringCalculations() (line 822) that will catch if the friction coefficients are improperly set to 0.0.