ExponentialSpringForce: Constructor arguments now named with monogram notation.
In the constructor for ExponentialSpringForce, I replaced "contactPlaneTransform" with "X_GP". Not only is X_GP shorter, it also gives the user more information about what the transform represents and does when used as an operator.

I also added the suffix "_B" to constructor arguments "body" and "station" to be crystal clear about frame.  The argument names are now body_B and station_B.  The suffixes were added through out the code, as well, including in ExponentialSpringForceImpl.  This change tightens up the code conceptually, and I think will make the code easier to follow.

Finally, comments relating to the above variable name changes were edited/enhanced to clarify issues of frame.