Improved Sliding algorithm for ExponentialSpringForce (#746)
* Improved Sliding algorithm for ExponentialSpringForce

This commit includes:

1. A new algorithm for handling the Sliding State.

2. A modification to the API that I believe will be more intuitive for the user. Specifically, the Friction Spring Zero is now consistently referred to as the Elastic Anchor Point (a term first suggested by Michael Sherman).

3. Improved initial conditions and comments in the Test_Adhoc - ExponentialSpringsComparison program.

For details refer to the description given in the Pull Request.

* ExponentialSpringForce: Constructor arguments now named with monogram notation.

In the constructor for ExponentialSpringForce, I replaced "contactPlaneTransform" with "X_GP". Not only is X_GP shorter, it also gives the user more information about what the transform represents and does when used as an operator.

I also added the suffix "_B" to constructor arguments "body" and "station" to be crystal clear about frame.  The argument names are now body_B and station_B.  The suffixes were added through out the code, as well, including in ExponentialSpringForceImpl.  This change tightens up the code conceptually, and I think will make the code easier to follow.

Finally, comments relating to the above variable name changes were edited/enhanced to clarify issues of frame.

Co-authored-by: Clay Anderson <fcanderson@users.noreply.github.com>