Replaced sprintf() with snprintf() in String.h
The Mac build fails because of the use of sprintf(), which has been deprecated.  Instead, snprint() should be used.

In snprintf(), writing into the character buffer is limited to "n" bytes and therefore, if n is chosen correctly, prevents writing past the end of the buffer and closes a security hole.