Added access to State indices for ExponentialSpringForce
Class ExponentialSpringForce has 4 states:
- Static coefficient of friction (discrete)
- Kinetic coefficient of friction (discrete)
- Sliding state (auto-update discrete)
- Elastic anchor point (auto-update discrete)

A typical user will not need the State indices of these states.

Access to the indices is provided so that other modeling frameworks written on top of Simbody, such as OpenSim, can get the values of a States in a generic way.  Specifically, with the indices, the values can be obtained by calling

SimTK::Subsystem::getDiscreteVariable(const State& s, DiscreteVariableIndex index)

A few trailing white spaces were also removed.