Added access to State indices for ExponentialSpringForce (#763)
* Added access to State indices for ExponentialSpringForce

Class ExponentialSpringForce has 4 states:
- Static coefficient of friction (discrete)
- Kinetic coefficient of friction (discrete)
- Sliding state (auto-update discrete)
- Elastic anchor point (auto-update discrete)

A typical user will not need the State indices of these states.

Access to the indices is provided so that other modeling frameworks written on top of Simbody, such as OpenSim, can get the values of a States in a generic way.  Specifically, with the indices, the values can be obtained by calling

SimTK::Subsystem::getDiscreteVariable(const State& s, DiscreteVariableIndex index)

A few trailing white spaces were also removed.

* Testing index accessors for exponential springs

The unit tests in TestExponentialSpring.cpp now include tests for
- getMuStaticStateIndex()
- getMuKineticStateIndex()
- getSlidingStateIndex()
- getAnchorPointStateIndex()

* Shortened documentation for state index accessors.

* Revamped documentation for state index accessors.

All state index accessors are now located at the end of ExponentialSpringForce.h and are collected under a Doxygen group named "State Indices (Advanced)". As a result, some repetition has been eliminated, and the clarity of the documentation improved.

---------

Co-authored-by: Clay Anderson <fcanderson@users.noreply.github.com>