Added precision argument to String(const T&)
Added an optional parameter, 'precision', to the templatized String constructor. It specifies the number of significant figures used to convert a numerical type to a String. The default value of the 'precision' is 6, which is typically the default precision of std::ostream.

I also updated the Doxygen documentation for this constructor.