Added an optional precision argument
I added an optional 'precision' parameter to three templatized methods:
- String::String<T>()
- Xml::Element::Element<T>()
- Xml::Element::setValueAs<T>()

To support the default value, I added the following constant:
String::DefaultOutputPrecision{6}

A value of 6 was chosen to agree with the default precision of std::ostream and so that the behavior of these methods after the additions will be the same as it was prior.

I revised the documentation for the above methods and the constant.