XML String Precision (#776)
* Added precision argument to String(const T&)

Added an optional parameter, 'precision', to the templatized String constructor. It specifies the number of significant figures used to convert a numerical type to a String. The default value of the 'precision' is 6, which is typically the default precision of std::ostream.

I also updated the Doxygen documentation for this constructor.

* Added an optional precision argument

I added an optional 'precision' parameter to three templatized methods:
- String::String<T>()
- Xml::Element::Element<T>()
- Xml::Element::setValueAs<T>()

To support the default value, I added the following constant:
String::DefaultOutputPrecision{6}

A value of 6 was chosen to agree with the default precision of std::ostream and so that the behavior of these methods after the additions will be the same as it was prior.

I revised the documentation for the above methods and the constant.

* Added a unit test for output precision.

To test the ability to modify the number of significant figures written to XML output, I added the subtest 'testOutputPrecision' to testXml.cpp.

In addition to Xml::Element::Element<T>() and Xml::Element::setValueAs<T>(), the subtest also tests String::String<T>(), which is the low-level templatized String constructor where precision control is achieved.

* Typo corrections and tweaks.

I proofed the documentation and made a few minor changes.

* More tweaks.

Just some developer notes.

* Simplified and refined developer notes.

I removed any reference to "API Adjustments" but retained the list of methods where the constant String::DefaultOutputPrecision is used.

* Corrected typos, removed dead code, improved name

"n" --> "num_precisions"