CableSpan: Via points (#828)
This change adds support for including via points in a CableSpan's path. Via points can be added to a CablePath by specifying a mobilized body index and a body-fixed station that define the position of the via point in the multibody system. Similarly to obstacles, via points are handled in the order that they are added to the CableSpan. Internally, via points divide the cable into segments, and a separate path optimization is performed for each segment.