bring drawing menus and sliders into the scene lock
It can happen that during the rendering of the menus or the sliders a
new command is for adding another menu is received. In this case, the
`menus` vector can become reallocated, making the current object
invalid.

This commit puts the scene lock around rendering the menus and sliders,
to ensure that either a menu is added or drawn.